Testing gpu

+++ examples/Test/verify.go
@@ -7,13 +7,13 @@ import (
 	"testing"
 )
 
-func VerifyScreen(t *testing.T, testname string, w int, h int, setup bool) {
-	err := gpu.CaptureToFile("./test-outputs/"+testname+".png", 0, 0, w, h)
+func VerifyScreen(t *testing.T, testname string, w float32, h float32, setup bool) {
+	err := gpu.CaptureToFile("./test-outputs/"+testname+".png", 0, 0, int(w), int(h))
 	if err != nil {
 		slog.Error("Capture to file failed, ", "file", "test-outputs/"+testname+".png", "error", err.Error())
 	}
 	if setup {
-		err = gpu.CaptureToFile("./test-assets/"+testname+".png", 0, 0, w, h)
+		err = gpu.CaptureToFile("./test-assets/"+testname+".png", 0, 0, int(w), int(h))
 	}
 	img1, err := gpu.LoadImage("./test-assets/" + testname + ".png")
 	if err != nil {
